- Fixed fnv1a hashing - Implemented updated (1.2.0) hashing

+++ WolvenKit.Common/FNV1A/FNV1A64HashAlgorithm.cs
@@ -78,7 +78,7 @@ namespace WolvenKit.Common.FNV1A
             {
                 unchecked
                 {
-                    hash = (hash ^ (uint)b) * (uint)FnvHashPrime;
+                    hash = (hash ^ (uint)b) * FnvHashPrime;
                 }
             }
